pl2: moved palettemapviewer into its own package
